Solucionado bug, por problemas entre int y double

+++ src/Vertex.cpp
@@ -5,7 +5,7 @@ Vertex::Vertex() : m_x(0), m_y(0)
 {
 }
 
-Vertex::Vertex(int x, int y) : m_x(x), m_y(y)
+Vertex::Vertex(double x, double y) : m_x(x), m_y(y)
 {
 }
